+++ demo/app/models/group.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class Group < ApplicationEntity
+  component Name
+  component Description
+end